refactor(web-ui): change project structure
