--- packages/frontend/web-ui/src/CornieApp.jsx
@@ -1,13 +1,0 @@
-import { useState } from 'react';
-import './App.css';
-import { Login } from './components/Login';
-
-function CornieApp() {
-  return (
-    <>
-      <Login />
-    </>
-  )
-}
-
-export default CornieApp
