+++ packages/frontend/web-ui/src/cornie/pages/CornieApp.jsx
@@ -1,0 +1,12 @@
+import '../../scss/layout/App.css';
+import { Login } from '../../auth/pages/Login';
+
+function CornieApp() {
+  return (
+    <>
+      <Login />
+    </>
+  )
+}
+
+export default CornieApp;
