+++ packages/frontend/web-ui/src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import CornieApp from './CornieApp';
-import './index.css';
+import CornieApp from './cornie/pages/CornieApp';
+import './scss/layout/index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
