+++ packages/frontend/web-ui/src/scss/layout/index.css
